feat: add spinner when loading
